+++ packages/world/mud.config.ts
@@ -4,7 +4,7 @@ export default mudConfig({
   worldImportPath: "../../",
   worldgenDirectory: "interfaces",
   worldInterfaceName: "IBaseWorld",
-  namespace: "world", // NOTE: this namespace is only used for tables, the core system is deployed in the root namespace.
+  namespace: "world" as const, // NOTE: this namespace is only used for tables, the core system is deployed in the root namespace.
   userTypes: {
     ResourceId: { filePath: "@latticexyz/store/src/ResourceId.sol", internalType: "bytes32" },
   },