+++ packages/recs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.0.0-next.13
+
+### Patch Changes
+
+- Updated dependencies [52182f70]
+  - @latticexyz/utils@2.0.0-next.13
+  - @latticexyz/schema-type@2.0.0-next.13
+
 ## 2.0.0-next.12
 
 ### Patch Changes