+++ packages/std-contracts/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @latticexyz/std-contracts
 
+## 2.0.0-next.13
+
 ## 2.0.0-next.12
 
 ## 2.0.0-next.11